Big hairy audacious packaging refactor 👹
* Integrated poetry.
* Integrated autohooks.
* Reformatted everything with black.
* Added pytest.
