+++ .github/workflows/build.yml
@@ -23,14 +23,12 @@ jobs:
         uses: actions/setup-python@v4
         with:
           # https://github.com/actions/setup-python#caching-packages-dependencies
-          cache: pip
+          cache: poetry
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .
+        run: poetry install --include=test
       - name: Run tests
-        run: python -m unittest
+        run: poetry run pytest
 
   docs:
     runs-on: ubuntu-latest
@@ -42,14 +40,12 @@ jobs:
         uses: actions/setup-python@v4
         with:
           # https://github.com/actions/setup-python#caching-packages-dependencies
-          cache: pip
-          python-version: "3.12"
+          cache: poetry
+          python-version: ${{ matrix.python-version }}
       - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .[docs-builder]
+        run: poetry install --include=dev
       - name: Check docs build
         run: |
           cd docs
           mkdir -p _static
-          make html
+          poetry run make html
