--- MANIFEST.in
@@ -1,3 +1,0 @@
-include LICENCE.txt
-graft test
-prune **/__pycache__
