+++ class_registry/__init__.py
@@ -1,5 +1,4 @@
-from .registry import *
-from .auto_register import *
-from .cache import *
-from .entry_points import *
-from .patcher import *
+__all__ = ["ClassRegistry", "RegistryKeyError"]
+
+from .base import RegistryKeyError
+from .registry import ClassRegistry
