+++ pyproject.toml
@@ -1,38 +1,65 @@
-# :see: https://peps.python.org/pep-0621/
+[build-system]
+requires = ["poetry-core=^1.7"]
+build-backend = "poetry.core.masonry.api"
 
-[project]
+# :see: https://greenbone.github.io/autohooks/installation.html
+[tool.autohooks]
+mode = "poetry"
+pre-commit = [
+    "autohooks.plugins.black",
+    "autohooks.plugins.ruff",
+    "autohooks.plugins.pytest"
+]
+
+# :see: https://python-poetry.org/docs/pyproject/
+[tool.poetry]
 name = "phx-class-registry"
 version = "4.1.0"
-description = "Factory+Registry pattern for Python classes"
+authors = ["Phoenix Zerin <phx@phx.nz>"]
+license = "MIT"
 readme = "README.rst"
-requires-python = ">= 3.10"
-license = { file = "LICENCE.txt" }
-authors = [
-    { email = "Phoenix Zerin <phx@phx.nz>" }
-]
+documentation = "https://class-registry.readthedocs.io/"
+repository = "https://github.com/todofixthis/class-registry"
+
+description = "Factory+Registry pattern for Python classes"
 keywords = [
     "design pattern",
     "factory pattern",
     "registry pattern",
+    "service registry",
 ]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Intended Audience :: Developers",
-    "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.12",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.10",
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 
-[project.optional-dependencies]
-"build-system" = ["build", "twine"]
-"docs-builder" = ["sphinx", "sphinx_rtd_theme"]
-"test-runner" = ["tox"]
+packages = [
+    { include = "class_registry" },
+]
+include = [
+    { path = "LICENCE.txt", format = "sdist" },
+    { path = "docs", format = "sdist" },
+    { path = "test", format = "sdist" }
+]
+exclude = ["docs/_build"]
 
-[project.urls]
-Documentation = "https://class-registry.readthedocs.io/"
+[tool.poetry.dependencies]
+python = "^3.10"
+
+[tool.poetry.group.dev.dependencies]
+autohooks = "^23.10.0"
+autohooks-plugin-black = "^23.10.0"
+autohooks-plugin-pytest = "^23.10.0"
+autohooks-plugin-ruff = "^23.11.0"
+pytest = "^7.4.3"
+sphinx = "^7.2.6"
+sphinx_rtd_theme = "^1.3.0"
+tox = "^4.11.3"
+
+[tool.poetry.group.test.dependencies]
+pytest = "^7.4.3"
+
+[tool.poetry.urls]
 Changelog = "https://github.com/todofixthis/class-registry/releases"
 Issues = "https://github.com/todofixthis/class-registry/issues"
-Repository = "https://github.com/todofixthis/class-registry"
