+++ tox.ini
@@ -1,5 +1,10 @@
+# :see: https://python-poetry.org/docs/faq/#usecase-3
 [tox]
 envlist = py3{12,11,10}
+isolated_build = true
 
 [testenv]
-commands = python -m unittest
+allowlist_externals = poetry
+commands_pre = poetry install --include=test
+commands = poetry run pytest
+skip_install = true
